+++ src/helper/switch.js
@@ -1,6 +1,9 @@
 import React from 'react';
+import TeamCard from '../components/browse/TeamCard';
 
-export const userTeamSwitch = (teams=[], dev=false, client=false) => {
+import { teamListConfig } from './teams';
+
+export const userTeamSwitch = (teams=[], dev=false, client=false, devs, clients) => {
   if(dev && dev.id) {
       return(
         <>
@@ -28,19 +31,17 @@ export const userTeamSwitch = (teams=[], dev=false, client=false) => {
         </>
       )
     } else if (client) {
+  
+      const clientTeams = teams.filter(team => team.c_id === client.id);
+
+      const clientList = teamListConfig(clientTeams, devs, [ client ]);
+
       return (
         <>
           <h2>{client.first_name}'s Results:</h2>
           {
-            teams.map(team => {
-              switch(client.id) {
-                case team.c_id:
-                  return <div><p>{team.project_title}</p></div>
-    
-                default:
-                  return null
-              }
-            })
+            teams.length ? clientList.map(team => <TeamCard team={team} />)
+                         : null
           }
         </>
       )
