+++ src/views/browse/BrowseTeams.js
@@ -3,50 +3,14 @@ import TeamCard from '../../components/browse/TeamCard';
 
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
+import { teamListConfig } from '../../helper/teams';
 
 const BrowseTeams = props => {
   const devs = useSelector(state => state.clientReducer.allDevs);
   const clients = useSelector(state => state.devReducer.allClients);
   const teams = useSelector(state => state.teamReducer.allTeams);
   
-  const teamList = teams.map(team => {
-    const detailed = { ...team }
-
-    clients.map(client => {
-      if( detailed.c_id === client.id ) {
-        delete detailed.c_id;
-        delete detailed.client_first;
-        delete detailed.client_last;
-        detailed.client = client;
-      }
-    });
-
-    devs.map(dev => {
-
-      if(dev.id === detailed.b_id) {
-        delete detailed.b_id;
-        detailed.backend = dev;
-      }
-
-      if(dev.id === detailed.f_id) {
-        delete detailed.f_id;
-        detailed.frontend = dev;
-      }
-
-      if(dev.id === detailed.u_id) {
-        delete detailed.u_id;
-        detailed.ui = dev;
-      }
-
-      if(dev.id === detailed.d_id) {
-        delete detailed.d_id;
-        detailed.devops = dev;
-      }
-
-    });
-
-    return detailed;
-  });
+  const teamList = teamListConfig(teams, devs, clients);
   
   return(
     <div>
@@ -58,7 +22,7 @@ const BrowseTeams = props => {
                                                   </Link>
       }
 
-      <h1>Past Teams:</h1>
+      <h1>All Teams:</h1>
       {
         teamList.map(team => {
           return(
